requirements.txt
et_xmlfile==2.0.0
iniconfig==2.1.0
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
tzdata==2025.2
# App Dependencies (Migrated from ai4scribe_old)
fastapi
uvicorn
jinja2
google-generativeai
faster-whisper
